tmux layouts updates

+++ tmux/layouts.sh
@@ -1,15 +1,17 @@
 #!/usr/bin/env bash
 # This script is used to load a tmux layout into the current session
 # Tmux layouts are scripts that execute tmux commands to create a series of windows
+# The first argument, if provided as "tns" indicates that this is being run from a call to "tns" which creates a new session
 
-LAYOUT_DIR=~/.tmux-layouts/
+# tlayout - tmux layout
+layout_dir=~/.tmux-layouts/
 
-if [[ ! -d "$LAYOUT_DIR" ]]; then
-    echo "$LAYOUT_DIR does not exist" >&2
+if [[ ! -d "$layout_dir" ]]; then
+    echo "$layout_dir does not exist" >&2
     exit 1
 fi
 
-layout=$(find $LAYOUT_DIR -type f -name '*.sh' \
+layout=$(find $layout_dir -type f -name '*.sh' \
     | awk -F'/' 'BEGIN { print "default" } { sub(/\.sh$/, ""); print $NF }' \
     | gum filter --limit=1 --indicator=">")
 
@@ -23,8 +25,13 @@ if [[ "$layout" == "default" ]]; then
     exit 0
 fi
 
-/$LAYOUT_DIR/${layout}.sh
+/$layout_dir/${layout}.sh
 
 # We need to handle the original window for the new session
 tmux select-window -t :2
-tmux kill-window -t :1
+
+# If we run this script directly from an existing session then we want to kill the first window.
+# Otherwise, when running from "tns" this script will exit and kill the first window.
+if [[ "$1" != "tns" ]]; then
+    tmux kill-window -t :1
+fi
